+++ src/main/java/bo/edu/uto/encuestadea/dominios/AccesoUsuario.java
@@ -2,24 +2,24 @@ package bo.edu.uto.encuestadea.dominios;
 
 public class AccesoUsuario implements java.io.Serializable {
 
-	private Long id_unidad_usuario;
-	private Long id_usuario;
+	private Integer id_unidad_usuario;
+	private Integer id_usuario;
 	private String id_facultad;
-	private Long id_unidad;
+	private Integer id_unidad;
 
-	public Long getId_unidad_usuario() {
+	public Integer getId_unidad_usuario() {
 		return id_unidad_usuario;
 	}
 
-	public void setId_unidad_usuario(Long id_unidad_usuario) {
+	public void setId_unidad_usuario(Integer id_unidad_usuario) {
 		this.id_unidad_usuario = id_unidad_usuario;
 	}
 
-	public Long getId_usuario() {
+	public Integer getId_usuario() {
 		return id_usuario;
 	}
 
-	public void setId_usuario(Long id_usuario) {
+	public void setId_usuario(Integer id_usuario) {
 		this.id_usuario = id_usuario;
 	}
 
@@ -31,11 +31,11 @@ public class AccesoUsuario implements java.io.Serializable {
 		this.id_facultad = id_facultad;
 	}
 
-	public Long getId_unidad() {
+	public Integer getId_unidad() {
 		return id_unidad;
 	}
 
-	public void setId_unidad(Long id_unidad) {
+	public void setId_unidad(Integer id_unidad) {
 		this.id_unidad = id_unidad;
 	}
