+++ src/main/java/bo/edu/uto/encuestadea/mapas/IntegranteComisionMapa.java
@@ -15,7 +15,7 @@ public interface IntegranteComisionMapa {
 
 	List<IntegranteComision> getIntegrantesComision(Unidad unidad) throws DataAccessException;
 	IntegranteComision getIntegranteComision(IntegranteComision integranteComision)throws DataAccessException;
-	List<IntegranteComision> getIntegrantesComisionPorIdUsuario(Long id_usuario) throws DataAccessException;
+	List<IntegranteComision> getIntegrantesComisionPorIdUsuario(Integer id_usuario) throws DataAccessException;
 	Integer insertarIntegrante(IntegranteComision ic) throws DataAccessException;
 	int borrarIntegrante(IntegranteComision ic)throws DataAccessException;
